requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
behave==1.2.6
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
easyparse==1.0.6
et_xmlfile==2.0.0
Flask==3.1.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
homepage==0.5.7
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.2
parse_type==0.6.4
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
robotframework==7.2.2
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.7.1
selenium==4.28.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yt-dlp==2025.2.19
zope.event==5.0
zope.interface==7.2
